Add support for SOA/NS record creation during zone creation
In order to produce a valid zone, the SOA record must be
properly constructed, and NS records must be created to
provide delegations to the nameservers for the zone.

The new 'soa' section of the 'properties' parameter
allows the various SOA fields to be populated, and the
existing 'nameservers' parameter is used to create the
NS records.

The only required configuration changes are the 'mname'
and 'rname' parameters in the 'soa' section.