Improve 'changed' reporting of rrsets module
When the 'rrsets' module does not actually need to make any changes,
it should not issue the zone-modification API call or report a
'changed' status. The existing code partially attempted to implement
this optimization, but it did not work unless 'keep' was set to True,
and the comparison between the provided and existing lists of records
was ineffective because extraneous attributes were present in the
'existing' record list. Finally, the comparison of record lists did
not ensure that they were sorted before comparison.